helper/no-ordered-equal.ts
export function noOrderedDeepEqual(source: any, target: any): boolean {
  if(typeof source !== typeof target) return false;
  if(Number.isNaN(source)) return isNaN(target);

  // for undefined/number/string/symbol/boolean/null/symbol/function
  if(typeof source !== 'object' || source === null || target === null) return source === target;

  // for array
  if(source instanceof Array && target instanceof Array) {
    if(source.length !== target.length) return false;
 
    return source.every((prop) => {
      return target.some(obj => noOrderedDeepEqual(prop, obj));
    });
  }

  // for Object
  let props = Object.keys(source);
  return props.every((prop) => noOrderedDeepEqual(source[prop], target[prop]));
}